Profit calculating logic is modified & functional test case (2) is modified.

+++ src/main/java/lotto/Application.java
@@ -236,7 +236,7 @@ public class Application {
     }
 
     private static void printProfit(int buy_price, int total) {
-        float profit = ((float) total / buy_price) * 100;
+        double profit = ((double) total / buy_price) * 100;
 
         System.out.printf("총 수익률은 %,.1f%%입니다.", profit);
     }
